components reorganised within atomic structure to improve readability and maintainability
