+++ src/atomic/molecules/orderSubmitted/OrderSubmitted.component.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import './OrderSubmitted.styles.scss';
+
+const OrderSubmitted: React.FC = () => {
+	return <div className='order-submitted'>Thanks, weve got your order</div>;
+};
+
+export default OrderSubmitted;
